almost there

+++ include/methods.inl
@@ -70,7 +70,7 @@ void randomFill(std::vector<Obj> &v, const Obj l, const Obj u, const unsigned in
         v[i] = distr(eng);
 }
 
-long double calculateTime(const std::vector<long int> v, int (*function)(std::vector<long int>, long int, int, int ), long int x, int l, int r ){
+long double calculateTime(const std::vector<long int> v, int (*function)(std::vector<long int>, long int, int, int ), long int x, long int l, long int r ){
 
     long double duration = 0;
     for (int i = 0; i < 99; ++i)
